cloud_functions/invoice_generator/requirements.txt
reportlab
supabase-py